copied parsing files from main.c to parsing.c
